refactor(macros): encapsulate code into other module
